Cleaning up (#767)
* Kill playerAPI

* Gut more events and compat that is now obsolete

* Remove commented out code

* Remove final modifier from methods

* Make more stuff final

* Remove slow building ring

* Protected -> private in final classes

* More cleaning

* More cleaning v2

* Purging

* Clean DevHelper

* Clean DevHelper 2

* delete DevHelper

* remove useless IFMLLoadingPlugin.MCVersion annotation from @Mod files

* check for obfuscation only once in the IFMLloadingPlugin.injectData

* don't instantiate the static class ASMConfig

* delete unused and empty implementation of IFMLCallHook

* delete empty class

* delete ClassesToTransform class that just holds the class names constants

* delete unused methods and unsless logging in dummymod container

* delete unused transformer ClassTransformer_TT_ThaumicRestorer

* spotless

* Clean

* Clean utils

* Build checkpoint

* Purge

* Stage 2

* Stage 3

* Stage 4

* Stage 5

* Stage 6

* Spotless

* Imports

* Stage idk

* Stage 1

* Fix issue

* Spotless

* Format numbers

---------

+++ src/main/java/gtPlusPlus/everglades/block/BlockDarkWorldSludgeFluid.java
@@ -11,8 +11,6 @@ public class BlockDarkWorldSludgeFluid extends Fluid {
 
     protected static int mapColor = 0xFFFFFFFF;
     protected static float overlayAlpha = 0.2F;
-    // protected static SoundEvent emptySound = SoundEvents.ITEM_BUCKET_EMPTY;
-    // protected static SoundEvent fillSound = SoundEvents.ITEM_BUCKET_FILL;
     protected static Material material = SLUDGE;
 
     public BlockDarkWorldSludgeFluid(String fluidName, int rgbColour) {
@@ -23,7 +21,7 @@ public class BlockDarkWorldSludgeFluid extends Fluid {
         super(fluidName);
         setColor(rgbColour);
         if (overlayAlpha != null) {
-            setAlpha(overlayAlpha.floatValue());
+            setAlpha(overlayAlpha);
         } else {
             setAlpha(0);
         }
